% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_simplifyhelper.futureplanprint.R
\name{simplifyhelper.futureplanprint}
\alias{simplifyhelper.futureplanprint}
\title{Helper function to create a character of the future::plan object}
\usage{
simplifyhelper.futureplanprint(future_plan)
}
\arguments{
\item{future_plan}{future::plan object to be converted to character string for printing}
}
\value{
Printable plan of future
}
\description{
Helper function to create a character of the future::plan object
}
\author{
Mayuresh Kinare
}
\keyword{plan}
