% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_simplifyhelper.handlemultisession.R
\name{simplifyhelper.handlemultisession}
\alias{simplifyhelper.handlemultisession}
\title{Helper function to handle future::plan object}
\usage{
simplifyhelper.handlemultisession(future_plan)
}
\arguments{
\item{future_plan}{future::plan object to be scrubbed off the multisession object}
}
\value{
Scrubbed future::plam object
}
\description{
Helper function to handle future::plan object
}
\author{
Mayuresh Kinare
}
\keyword{multisession}
\keyword{plan,}
