% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_simplifyhelper.listtodata.R
\name{simplifyhelper.listtodata}
\alias{simplifyhelper.listtodata}
\title{Helper function to create a data table from a vector of columns}
\usage{
simplifyhelper.listtodata(column_y, object_x)
}
\arguments{
\item{column_y}{vector that need to be joined with the object_x}

\item{object_x}{An object to create the combination}
}
\value{
Table containing data table for all possible combinations of the column_y and object_x
}
\description{
Helper function to create a data table from a vector of columns
}
\author{
Mayuresh Kinare
}
\keyword{listtodata}
