% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_simplifyhelper.removesymbols.R
\name{simplifyhelper.removesymbols}
\alias{simplifyhelper.removesymbols}
\title{Helper function to remove certain symbols from the column names}
\usage{
simplifyhelper.removesymbols(data, columns, splits)
}
\arguments{
\item{data}{A data frame}

\item{columns}{A character vector of columns which to be stripped of the symbols, structure is c("Col1", "Col2", "Col3",.......)}

\item{splits}{A character vector of splits on which to be stripped of the symbols, structure is c("split_element1", "split_element2", "split_element3",.......)}
}
\value{
Data table with new column names stripped of the error causing symbols
}
\description{
Helper function to remove certain symbols from the column names
}
\author{
Mayuresh Kinare
}
\keyword{Remove}
\keyword{symbols}
