% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_simplifyhelper.splittodata.R
\name{simplifyhelper.splittodata}
\alias{simplifyhelper.splittodata}
\title{#' Helper function to create a data table from vector of split with contents of the split}
\usage{
simplifyhelper.splittodata(data, column_split)
}
\arguments{
\item{data}{The data table from which we have to create the split}

\item{column_split}{An column name which is used as a split column}
}
\value{
Table containing the object and the content of the object from the data table which was passed
}
\description{
#' Helper function to create a data table from vector of split with contents of the split
}
\author{
Mayuresh Kinare
}
\keyword{split,}
\keyword{splittodata}
